:art: (webpack) Finished express to micro transition
This commit completely moves us from Express to Micro for the dev server, including quiet output and
proper HMR support. Also removes the log-plugin class, because since micro flashes the port at you
when you launch and copies the URL to clipboard, the notification is unnecessary. Also, if the
normal port is busy, it'll fallback to another port, so we have no way to tell what port is really
being used.